SMTChecker experiments
+ `yarn check` runs SMTChecker on your contract
+ preprocessor step

To do anything useful, the checker relies on assert() statements in
code. While some of them might be OK to have in production code, most of
them use stack (because you need to explicitly store state sometimes).

To aid with that, there's a new preprocessor step that comments out all
lines with "#nonprod" suffix. Example:
https://rinkeby.etherscan.io/address/0xc0bBBd6ed9092Ed978CAbDAdC74cb021b8330d41#code
, look at withdraw() function at the end

`yarn check` is the only command that DOES NOT run the preprocessor and
has full access to all asserts().

It is not clear if the checked is useful, but it's worth playing around
with