SMTChecker experiments (#5)
* SMTChecker experiments

+ `yarn check` runs SMTChecker on your contract
+ preprocessor step

To do anything useful, the checker relies on assert() statements in
code. While some of them might be OK to have in production code, most of
them use stack (because you need to explicitly store state sometimes).

To aid with that, there's a new preprocessor step that comments out all
lines with "#nonprod" suffix. Example:
https://rinkeby.etherscan.io/address/0xc0bBBd6ed9092Ed978CAbDAdC74cb021b8330d41#code
, look at withdraw() function at the end

`yarn check` is the only command that DOES NOT run the preprocessor and
has full access to all asserts().

It is not clear if the checked is useful, but it's worth playing around
with


BadContract has bad design in `withdraw` method: external action
followed by state modification.

SMTChecker catches that! Moreover, it correctly proves that
`nonReentrant` modifier fixes the issue.

We can't use `NonReentrant` from OpenZeppelin - they use `uint256
_status` as the 'entered/not entered' status and that blows up CHC's
exploration space. There's no good reason to use uint256 (except maybe
variables packing?), so a home-grown version of `nonReentrant` with a
simple boolean field makes the trick